[smb] SMBexec: avoid looping by adding numbers of tires.

+++ cme/protocols/smb/smbexec.py
@@ -26,7 +26,8 @@ class SMBEXEC:
         hashes=None,
         share=None,
         port=445,
-        logger=cme_logger
+        logger=cme_logger,
+        tires=None
     ):
         self.__host = host
         self.__share_name = "C$"
@@ -51,6 +52,7 @@ class SMBEXEC:
         self.__aesKey = aesKey
         self.__doKerberos = doKerberos
         self.__kdcHost = kdcHost
+        self.__tires = tires
         self.logger = logger
 
         if hashes is not None:
@@ -150,21 +152,26 @@ class SMBEXEC:
         if self.__retOutput is False:
             self.__outputBuffer = ""
             return
+        tires = 0
         while True:
             try:
+                cme_logger.info(f"Attempting to read {self.__share}\\{self.__output}")
                 self.__smbconnection.getFile(self.__share, self.__output, self.output_callback)
                 break
             except Exception as e:
-                print(e)
-                if str(e).find("STATUS_SHARING_VIOLATION") >= 0:
+                if tires >= self.__tires:
+                    self.logger.fail(f'SMBEXEC: Get output file error, maybe go detection by AV software, please try "--get-output-tires" option. If it\'s still failing maybe something is blocking the schedule job, try another exec method')
+                    break
+                if str(e).find("STATUS_SHARING_VIOLATION") >= 0 or str(e).find("STATUS_OBJECT_NAME_NOT_FOUND") >= 0:
                     # Output not finished, let's wait
                     sleep(2)
-                    pass
+                    tires += 1
                 else:
-                    self.logger.debug(e)
-                    pass
-
-        self.__smbconnection.deleteFile(self.__share, self.__output)
+                    raise
+        
+        if self.__outputBuffer:
+            cme_logger.debug(f"Deleting file {self.__share}\\{self.__output}")
+            self.__smbconnection.deleteFile(self.__share, self.__output)
 
     def execute_fileless(self, data):
         self.__output = gen_random_string(6)
